Implement random-word for windows
Should increase the size of the executable by a metric shit ton, but hey,
at least it's still an option.

Need to finish up some tests for this, break out some of the functionality into
sub-packages, give it a real name and put it in its own separate repo. Not bad
for a long weekend's worth of work given that I started this in the grasp of
a wild fever dream.